ENH: plate reader analysis improvements
* Related to #28.

* The main image analysis module was using the
older function name `plot_input_data_cesar_CG`,
which has been updated to the new
`plot_input_data_cesar_MD`.

* Add a `joblib` cache decoration for `skimage_hough_transform()`;
it now runs in 2 seconds instead of 2 minutes when repeated. The purpose
is to improve image analysis iteration times when working on
other parts of the code.

* Minor `.gitignore` update to deal with more Python artifacts.

* Early draft of Determinant of Hessian (DoH) approach to measuring
the blobs/bubbles in the plate reader images.

* `plot_input_data_cesar_MD()` was adjusted to allow a few
more customizations that proved useful in this expanded blob
detection work, including log scaling and labeling of the color
bar/scheme.