CI: add initial CI testing
* This adds linting and unit testing based on the
in-house GitLab CI we've been using at
https://lisdi-git.lanl.gov/ldrd_dr_neat/ldrd_neat_ml. I've
excluded the "functional test" we were previously using
in-house because it times out after running for an hour,
and is not really useful at this time.

* This CI configuration runs with more Python versions and
platforms (MacOS and Linux) compared to the in-house CI
config, so is a bit more thorough in that sense.